DEL: Remove `SymbolMap`
